Rename props.tsx to common-props.tsx
