+++ src/lib/seam/components/AccessCodeDetails/AccessCodeDetails.tsx
@@ -16,7 +16,7 @@ import { DeviceDetails } from 'lib/seam/components/DeviceDetails/DeviceDetails.j
 import {
   type CommonProps,
   withRequiredCommonProps,
-} from 'lib/seam/components/props.js'
+} from 'lib/seam/components/common-props.js'
 import { Alerts } from 'lib/ui/Alert/Alerts.js'
 import { Button } from 'lib/ui/Button.js'
 import { copyToClipboard } from 'lib/ui/clipboard.js'
