+++ src/lib/seam/components/AccessCodeTable/AccessCodeTable.tsx
@@ -18,7 +18,7 @@ import { EditAccessCodeForm } from 'lib/seam/components/EditAccessCodeForm/EditA
 import {
   type CommonProps,
   withRequiredCommonProps,
-} from 'lib/seam/components/props.js'
+} from 'lib/seam/components/common-props.js'
 import { IconButton } from 'lib/ui/IconButton.js'
 import { ContentHeader } from 'lib/ui/layout/ContentHeader.js'
 import { EmptyPlaceholder } from 'lib/ui/Table/EmptyPlaceholder.js'
