+++ src/lib/seam/components/common-props.tsx
